requirements.txt
telegram == 0.0.1
python-telegram-bot == 13.7

mysql_connector_python == 8.0.26

wikipedia == 1.4.0
psutil == 5.8.0
speedtest_cli == 2.1.3